+++ src/onegov.org/onegov/org/app.py
@@ -36,6 +36,7 @@ class OrgApp(Framework, LibresIntegration, ElasticsearchApp, MapboxApp,
     export = directive(directives.ExportAction)
     userlinks = directive(directives.UserlinkAction)
     directory_search_widget = directive(directives.DirectorySearchWidgetAction)
+    settings_view = directive(directives.SettingsView)
 
     #: the version of this application (do not change manually!)
     version = '0.46.5'